+++ inst/include/cpp11/strings.hpp
@@ -172,8 +172,12 @@ typedef r_vector<r_string> strings;
 template <typename T>
 inline void r_vector<T>::push_back(const named_arg& value) {
   push_back(value.value());
+  if (Rf_xlength(names()) == 0) {
+    cpp11::writable::strings new_nms(size());
+    names() = new_nms;
+  }
   cpp11::writable::strings nms(names());
-  nms.push_back(value.name());
+  nms[size() - 1] = value.name();
 }
 
 }  // namespace writable